job-portal/backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.2.0
psycopg2-binary
django-cors-headers
